URE CyberArk 1.0.0-beta - Initial release (#160)
CyberArk plugin performs action on CyberArk users like - Add users to a role and Remove users from a role on the basis of business rule. 

The actions will only be performed if the users matched in the business rule are present on the CyberArk platform.

Note: It does not support fetching users/hosts or their scores from CyberArk.